Change interval value to int and finish the patch of intervals in the frontend

+++ lib/smart_garden/schemas/interval.ex
@@ -7,7 +7,7 @@ defmodule SmartGarden.Interval do
 
   schema "intervals" do
     field :name, :string
-    field :value, :float
+    field :value, :integer
     field :action, :string
     field :execution_schedule, :string
     belongs_to :device, SmartGarden.Device
